feat: moved generation configuration to core
