+++ core/src/main/java/org/jboss/sbomer/core/config/DefaultProcessingConfig.java
@@ -15,23 +15,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.sbomer.cli.feature.sbom.config;
+package org.jboss.sbomer.core.config;
 
 import java.util.List;
 
-import jakarta.enterprise.context.ApplicationScoped;
-
 import org.jboss.sbomer.core.features.sbom.enums.ProcessorType;
 
-import io.smallrye.config.ConfigMapping;
 import io.smallrye.config.WithDefault;
 import io.smallrye.config.WithName;
 
 /**
  * @author Marek Goldmann
  */
-@ApplicationScoped
-@ConfigMapping(prefix = "sbomer.processing")
 public interface DefaultProcessingConfig {
     public interface ProcessorConfig {
         ProcessorType name();
