+++ core/src/main/resources/META-INF/sbomer-config.yaml
@@ -1,0 +1,37 @@
+#
+# JBoss, Home of Professional Open Source.
+# Copyright 2023 Red Hat, Inc., and individual contributors
+# as indicated by the @author tags.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+sbomer:
+  generation:
+    enabled: true
+    default-generator: maven-cyclonedx
+    generators:
+      maven-domino:
+        default-version: "0.0.97"
+        default-args: "--include-non-managed --warn-on-missing-scm"
+      maven-cyclonedx:
+        default-version: "2.7.9"
+        default-args: "--batch-mode"
+
+  processing:
+    enabled: true
+    auto-process: true
+    # List of default processors that will be run on every generated SBOM
+    default-processors:
+      - name: "default"
+      - name: "redhat-product"
