+++ core/src/main/resources/schemas/config.json
